update
Damnit.  It doesn't quite work.  Need to add some regex code to be sure that when each precinct runs the program only scrapes the data from the master csv for that precinct.  Right now stops at first number and matches to that.  Ok for townships with 9 or fewer precincts, but not for Algonquin, Dorr, Grafton, McHenry, or Nunda.